requirements.txt
Django==4.0.3
numpy==1.21.5
pandas==1.4.1
Pillow==9.0.1
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
tk==0.1.0
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tqdm==4.64.0

